+++ src/main/java/ru/itis/inf304/ConstructionLab12Sem2/WorkClasses/BeginningOfWork.java
@@ -11,7 +11,7 @@ public class BeginningOfWork extends AbstractWorkClass {
     }
 
     @Override
-    protected long getDuration() {
+    protected int getDuration() {
         return 0;
     }
