+++ src/main/java/ru/itis/inf304/ConstructionLab12Sem2/WorkClasses/CompletionOfConstruction.java
@@ -26,7 +26,7 @@ public class CompletionOfConstruction extends AbstractWorkClass {
     }
 
     @Override
-    protected long getDuration() {
+    protected int getDuration() {
         return 0;
     }
