+++ src/main/java/ru/itis/inf304/ConstructionLab12Sem2/WorkClasses/DoorInstallation.java
@@ -3,7 +3,7 @@ package ru.itis.inf304.ConstructionLab12Sem2.WorkClasses;
 import ru.itis.inf304.ConstructionLab12Sem2.AbstractWorkClass;
 
 public class DoorInstallation extends AbstractWorkClass {
-    long duration = 1;
+    int duration = 1;
     int dayOfBegin;
     int dayOfEnd;
 
@@ -29,7 +29,7 @@ public class DoorInstallation extends AbstractWorkClass {
     }
 
     @Override
-    protected long getDuration() {
+    protected int getDuration() {
         return duration;
     }
